fix(microapp): 添加json示例和接口文档地址
