+++ src/SKIT.FlurlHttpClient.ByteDance.MicroApp/SDK/AuthEntity/Models/AuthEntityActivateMerchantCodeResponse.cs
@@ -1,0 +1,7 @@
+namespace SKIT.FlurlHttpClient.ByteDance.MicroApp.SDK.AuthEntity.Models
+{
+    public class AuthEntityActivateMerchantCodeResponse : ByteDanceMicroAppAuthEntityResponse
+    {
+
+    }
+}
