+++ src/SKIT.FlurlHttpClient.ByteDance.MicroApp/SDK/AuthEntity/Models/AuthEntityGetAppidAuthResponse.cs
@@ -1,113 +1,7 @@
-using System;
 using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 
 namespace SKIT.FlurlHttpClient.ByteDance.MicroApp.SDK.AuthEntity.Models
 {
-    public class AuthEntityRoleBindOrUnRequest : ByteDanceMicroAppAuthEntityRequest
-    {
-        /// <summary>
-        /// 要绑定小程序的实体。（仅机构or服务商）
-        /// </summary>
-        [Newtonsoft.Json.JsonProperty("merchant_entity_id")]
-        [System.Text.Json.Serialization.JsonPropertyName("merchant_entity_id")]
-        public string MerchantEntityId { get; set; }
-
-        /// <summary>
-        /// 行业编码，泛知识10000
-        /// </summary>
-        [Newtonsoft.Json.JsonProperty("industry_code")]
-        [System.Text.Json.Serialization.JsonPropertyName("industry_code")]
-        public string IndustryCode { get; set; }
-
-        /// <summary>
-        /// 实体角色类型（1：老师 2：机构 3：服务商）
-        /// </summary>
-        [Newtonsoft.Json.JsonProperty("merchant_industry_role")]
-        [System.Text.Json.Serialization.JsonPropertyName("merchant_industry_role")]
-        public string MerchantIndustryRole { get; set; }
-
-    }
-
-    public class AuthEntityRoleBindOrUnResponse : ByteDanceMicroAppAuthEntityResponse
-    {
-
-    }
-
-    public class AuthEntityGetBindListRequest : ByteDanceMicroAppAuthEntityRequest
-    {
-
-    }
-    public class AuthEntityGetBindListResponse : ByteDanceMicroAppAuthEntityResponse
-    {
-        public static class Types
-        {
-            public class BindInfo
-            {
-                /// <summary>
-                /// 小程序appid
-                /// </summary>
-                [Newtonsoft.Json.JsonProperty("appid")]
-                [System.Text.Json.Serialization.JsonPropertyName("appid")]
-                public string AppId { get; set; }
-
-                /// <summary>
-                /// 角色id
-                /// </summary>
-                [Newtonsoft.Json.JsonProperty("role_id")]
-                [System.Text.Json.Serialization.JsonPropertyName("role_id")]
-                public string RoleId { get; set; }
-
-                /// <summary>
-                /// 1:个人(泛知识行业为老师),2:自营机构,3:服务商
-                /// </summary>
-                [Newtonsoft.Json.JsonProperty("industry_role")]
-                [System.Text.Json.Serialization.JsonPropertyName("industry_role")]
-                public int IndustryRole { get; set; }
-
-                /// <summary>
-                /// 1000:泛知识
-                /// </summary>
-                [Newtonsoft.Json.JsonProperty("industry_code")]
-                [System.Text.Json.Serialization.JsonPropertyName("industry_code")]
-                public long IndustryCode { get; set; }
-
-            }
-        }
-        /// <summary>
-        /// 小程序绑定的所有角色
-        /// </summary>
-        [Newtonsoft.Json.JsonProperty("bind_info_list")]
-        [System.Text.Json.Serialization.JsonPropertyName("bind_info_list")]
-        public List<Types.BindInfo> BindInfoList { get; set; }
-    }
-
-    public class AuthEntityGetAppidAuthRequest : ByteDanceMicroAppAuthEntityRequest
-    {
-        /// <summary>
-        /// 服务商实体ID
-        /// </summary>
-        [Newtonsoft.Json.JsonProperty("partner_entity_id")]
-        [System.Text.Json.Serialization.JsonPropertyName("partner_entity_id")]
-        public string PartnerEntityId { get; set; } = string.Empty;
-
-        /// <summary>
-        /// 被授权实体
-        /// </summary>
-        [Newtonsoft.Json.JsonProperty("merchant_entity_id")]
-        [System.Text.Json.Serialization.JsonPropertyName("merchant_entity_id")]
-        public string MerchantEntityId { get; set; } = string.Empty;
-
-        /// <summary>
-        /// 授权实体角色类型（1：老师 2：机构 ）
-        /// </summary>
-        [Newtonsoft.Json.JsonProperty("merchant_industry_role")]
-        [System.Text.Json.Serialization.JsonPropertyName("merchant_industry_role")]
-        public int MerchantIndustryRole { get; set; }
-    }
-
     public class AuthEntityGetAppidAuthResponse : ByteDanceMicroAppAuthEntityResponse
     {
         public static class Types
