+++ src/SKIT.FlurlHttpClient.ByteDance.MicroApp/SDK/AuthEntity/Models/AuthEntityRoleBindOrUnResponse.cs
@@ -1,0 +1,7 @@
+namespace SKIT.FlurlHttpClient.ByteDance.MicroApp.SDK.AuthEntity.Models
+{
+    public class AuthEntityRoleBindOrUnResponse : ByteDanceMicroAppAuthEntityResponse
+    {
+
+    }
+}
