+++ test/SKIT.FlurlHttpClient.ByteDance.MicroApp.UnitTests/ModelSamples/AuthEntity/AuthEntityActivateMerchantCodeRequest.json
@@ -1,0 +1,9 @@
+{
+    "appid": "ttxxxxxxxx",
+    "access_token": "xxxxxxxx",
+    "certificate_id": "100330xxxxxxx",
+    "industry_code": 10000,
+    "industry_role": 1,
+    "merchant_code": "99sxxxx",
+    "callback_url": ""
+}
