+++ test/SKIT.FlurlHttpClient.ByteDance.MicroApp.UnitTests/ModelSamples/AuthEntity/AuthEntityAddClassAuthResponse.json
@@ -1,0 +1,7 @@
+{
+    "class_auth_taskid": "Role_71790xxxxx",
+    "err": {
+        "err_code": 0,
+        "err_msg": ""
+    }
+}
