+++ test/SKIT.FlurlHttpClient.ByteDance.MicroApp.UnitTests/ModelSamples/AuthEntity/AuthEntityAddRoleResponse.json
@@ -1,0 +1,7 @@
+{
+    "add_role_taskid": "Role_717907890xxx4238252",
+    "err": {
+        "err_code": 0,
+        "err_msg": ""
+    }
+}
