+++ test/SKIT.FlurlHttpClient.ByteDance.MicroApp.UnitTests/ModelSamples/AuthEntity/AuthEntityAuthRoleRequest.json
@@ -1,0 +1,21 @@
+{
+    "appid": "tt427e34c3dd07777601",
+    "access_token": "080112184630xxx947774a45363371682f48762f4c67326974513d3d",
+    "partner_entity_id": "E_7208933888470106156",
+    "merchant_entity_id": "E_7208933889593688108",
+    "industry_code": 10000,
+    "auth_letters": [
+        {
+            "material_type": 1006,
+            "material_expiretime": "2033-01-01",
+            "material_paths": [
+                "certification/resource/c1bb104b7a84fb40e58f89a91cf19b22"
+            ]
+        }
+    ],
+    "scope": 2,
+    "nick": "哈哈哈",
+    "icon": "certification/resource/c1bb104b7a84fb40e58f89a91cf19b22",
+    "desc": "哈哈",
+    "merchant_industry_role": 2
+}
