+++ test/SKIT.FlurlHttpClient.ByteDance.MicroApp.UnitTests/ModelSamples/AuthEntity/AuthEntityAuthRoleResponse.json
@@ -1,0 +1,7 @@
+{
+    "bind_taskid": "AppidBind_721021xxx2534894636",
+    "err": {
+        "err_code": 0,
+        "err_msg": ""
+    }
+}
