+++ test/SKIT.FlurlHttpClient.ByteDance.MicroApp.UnitTests/ModelSamples/AuthEntity/AuthEntityByPartnerResponse.json
@@ -1,0 +1,9 @@
+{
+    "basic_auth_taskid": "Basic_7179078906xxx172716",
+    "class_auth_taskid": "Role_717907890xxx4238252",
+    "entity_id": "I_7179078xxx694139948",
+    "err": {
+        "err_code": 0,
+        "err_msg": ""
+    }
+}
