+++ test/SKIT.FlurlHttpClient.ByteDance.MicroApp.UnitTests/ModelSamples/AuthEntity/AuthEntityBySelfRequest.json
@@ -1,0 +1,83 @@
+{
+    "appid": "ttxxxxxxxx",
+    "access_token": "xxxxxxxx",
+    "basic_auth": {
+        "entity_type": 2,
+        "entity_name": "泛知识机构九",
+        "certificate_type": 2,
+        "enterprise": {
+            "certificate_id": "911120220xxx140399",
+            "certificate_materials": [
+                {
+                    "material_type": 1007,
+                    "material_expiretime": "2026-02-02",
+                    "material_paths": [
+                        "certification/resource/4c23f6490a4b65b1c537abdd765b1907"
+                    ]
+                }
+            ],
+            "legal_person": {
+                "name": "刘x国",
+                "id_number": "2102191959xxx06418",
+                "expire_time": "2028-03-03",
+                "front_path": "certification/resource/1f3ac1c1966a19d83e7607bfb3d38ce9",
+                "back_path": "certification/resource/11dd877fb32b439c4609f25f76ab05c1"
+            }
+        }
+    },
+    "class_auth": {
+        "industry_code": 10000,
+        "industry_class": {
+            "first_class": 70000,
+            "second_class": 70500,
+            "third_class": 0
+        },
+        "industry_role": 2,
+        "qualifications": [
+            {
+                "material_type": 1001,
+                "material_expiretime": "2030-04-04",
+                "material_paths": [
+                    "certification/resource/722460c5bca1bac5f76ba9ac10f5d030"
+                ]
+            }
+        ],
+        "institution": {
+            "record_name": "泛知识机构九",
+            "scene_type": "线上机构",
+            "subject_type": "企业工商户",
+            "logo_uri": "certification/resource/94699056669e84341b488570bdbe87d5",
+            "trademark_uri": "certification/resource/94699056669e84341b488570bdbe87d5",
+            "desc": "这是简介简介必须是不得少于12个字符",
+            "employee": {
+                "employee_material": {
+                    "name": "花x玲",
+                    "id_number": "32068219xxx8139122",
+                    "expire_time": "2030-05-05",
+                    "front_path": "certification/resource/cd29412b2392d0f1c488bd9aa1e650b7",
+                    "back_path": "certification/resource/d797a08e8d542cd5db0f84a1710900c0"
+                },
+                "cooperation_agreement": {
+                    "material_type": 1005,
+                    "material_expiretime": "2031-06-06",
+                    "material_paths": [
+                        "certification/resource/722460c5bca1bac5f76ba9ac10f5d030"
+                    ]
+                }
+            }
+        },
+        "partner": {
+            "company_type": "",
+            "company_addr": "",
+            "cooperation_cases": [
+                {
+                    "material_type": 1000,
+                    "material_expiretime": "",
+                    "material_paths": [
+                        ""
+                    ]
+                }
+            ]
+        }
+    }
+}
