+++ test/SKIT.FlurlHttpClient.ByteDance.MicroApp.UnitTests/ModelSamples/AuthEntity/AuthEntityBySelfResponse.json
@@ -1,0 +1,9 @@
+{
+    "basic_auth_taskid": "Basic_717907xxxxxx69324",
+    "class_auth_taskid": "Role_7xxxxxxx4860",
+    "entity_id": "E_717xxxxx556",
+    "err": {
+        "err_code": 0,
+        "err_msg": ""
+    }
+}
