+++ test/SKIT.FlurlHttpClient.ByteDance.MicroApp.UnitTests/ModelSamples/AuthEntity/AuthEntityGetAppidAuthRequest.json
@@ -1,0 +1,7 @@
+{
+    "appid": "tt427e34c3dd07777601",
+    "access_token": "0801121846307773635162617634556b772f72597545626155513d3d",
+    "partner_entity_id": "E_7208933888470106156",
+    "merchant_entity_id": "E_7208933889593688108",
+    "merchant_industry_role": 2
+}
