+++ test/SKIT.FlurlHttpClient.ByteDance.MicroApp.UnitTests/ModelSamples/AuthEntity/AuthEntityGetAppidAuthResponse.json
@@ -1,0 +1,31 @@
+{
+    "data": {
+        "appid": "tt427e34c3dd07777601",
+        "audit_auth_letter": {},
+        "auth_reason": "",
+        "auth_status": 2,
+        "desc": "哈哈",
+        "icon": "certification/resource/c1bb104b7a84fb40e58f89a91cf19b22",
+        "nick": "哈哈哈",
+        "online_auth_letter": {
+            "audit_reason": "",
+            "audit_status": 2,
+            "audit_taskid": "AppidBind_7209887453367517228",
+            "expire_time": "2033-01-01",
+            "qualifications": [
+                {
+                    "material_expiretime": "2033-01-01",
+                    "material_paths": [
+                        "certification/resource/c1bb104b7a84fb40e58f89a91cf19b22"
+                    ],
+                    "material_type": 1006
+                }
+            ]
+        },
+        "scope": 2
+    },
+    "err": {
+        "err_code": 0,
+        "err_msg": ""
+    }
+}
