+++ test/SKIT.FlurlHttpClient.ByteDance.MicroApp.UnitTests/ModelSamples/AuthEntity/AuthEntityGetAuditDetailRequest.json
@@ -1,0 +1,6 @@
+{
+    "appid": "ttxxxxx",
+    "access_token": "yyyyy",
+    "auth_type": 2,
+    "auth_taskid": "Role_7178xxxxxx53885996"
+}
