+++ test/SKIT.FlurlHttpClient.ByteDance.MicroApp.UnitTests/ModelSamples/AuthEntity/AuthEntityGetBasicAuthRequest.json
@@ -1,0 +1,5 @@
+{
+    "appid": "tta12xxxxxcbe01",
+    "access_token": "080yyyyyyy35494e62413d3d",
+    "entity_id": "E_7178703725698269228"
+}
