+++ test/SKIT.FlurlHttpClient.ByteDance.MicroApp.UnitTests/ModelSamples/AuthEntity/AuthEntityGetBasicAuthResponse.json
@@ -1,0 +1,55 @@
+{
+    "audit_qualification": {},
+    "auth_reason": "",
+    "auth_status": 2,
+    "basic_auth": {
+        "certificate_type": 2,
+        "enterprise": {
+            "certificate_id": "9111283673140399",
+            "certificate_materials": [
+                {
+                    "material_expiretime": "2028-02-02",
+                    "material_paths": [
+                        "certification/resource/4c2rewr0a4b65b1c537abdd765b1907"
+                    ],
+                    "material_type": 1007
+                }
+            ],
+            "legal_person": {
+                "back_path": "certification/resource/11dd87rewwt609f25f76ab05c1",
+                "expire_time": "2028-02-02",
+                "front_path": "certification/resource/1f3ac1c1966a1eteer7bfb3d38ce9",
+                "id_number": "320682198608139122",
+                "name": "花铃"
+            }
+        },
+        "entity_name": "泛知识机构三",
+        "entity_type": 2,
+        "individual": {}
+    },
+    "err": {
+        "err_code": 0,
+        "err_msg": ""
+    },
+    "online_qualification": {
+        "audit_reason": "",
+        "audit_status": 2,
+        "audit_taskid": "Basic_7178727377638260780",
+        "expire_time": "2028-02-02",
+        "qualifications": [
+            {
+                "material_expiretime": "2028-02-02",
+                "material_paths": ["certification/resource/4c23f6490aertgdsd5b1907"],
+                "material_type": 1007
+            },
+            {
+                "material_expiretime": "2028-02-02",
+                "material_paths": [
+                    "certification/resource/1f3ac1c196reras3e7607bfb3d38ce9",
+                    "certification/resource/11dd877r43rer4609f25f76ab05c1"
+                ],
+                "material_type": 1000
+            }
+        ]
+    }
+}
