+++ test/SKIT.FlurlHttpClient.ByteDance.MicroApp.UnitTests/ModelSamples/AuthEntity/AuthEntityGetBindListRequest.json
@@ -1,0 +1,2 @@
+{
+}
