+++ test/SKIT.FlurlHttpClient.ByteDance.MicroApp.UnitTests/ModelSamples/AuthEntity/AuthEntityGetClassAuthResponse.json
@@ -1,0 +1,35 @@
+{
+    "class_auths": [
+        {
+            "audit_qualification": {},
+            "industry_class": {
+                "first_class": 40000,
+                "second_class": 40600,
+                "third_class": 0
+            },
+            "industry_code": 10000,
+            "industry_role": 1,
+            "institution": {},
+            "online_qualification": {
+                "audit_reason": "",
+                "audit_status": 2,
+                "audit_taskid": "Role_7179078906694238252",
+                "expire_time": "2033-01-02",
+                "qualifications": [
+                    {
+                        "material_expiretime": "2033-01-02",
+                        "material_paths": [
+                            "certification/resource/01d86c188a9a6ba4e4bbbe3385ccf1f2"
+                        ],
+                        "material_type": 1001
+                    }
+                ]
+            },
+            "partner": {}
+        }
+    ],
+    "err": {
+        "err_code": 0,
+        "err_msg": ""
+    }
+}
