+++ test/SKIT.FlurlHttpClient.ByteDance.MicroApp.UnitTests/ModelSamples/AuthEntity/AuthEntityQueryEntityInfoRequest.json
@@ -1,0 +1,5 @@
+{
+    "certificate_id": "4111********1141858",
+    "access_token": "0801121846************************31513d3d",
+    "app_id": "tt81e********f39801"
+}
