+++ test/SKIT.FlurlHttpClient.ByteDance.MicroApp.UnitTests/ModelSamples/AuthEntity/AuthEntityQueryEntityInfoResponse.json
@@ -1,0 +1,7 @@
+{
+    "entity_id": "E_72062********437676",
+    "err": {
+        "err_code": 0,
+        "err_msg": "成功"
+    }
+}
