+++ test/SKIT.FlurlHttpClient.ByteDance.MicroApp.UnitTests/ModelSamples/AuthEntity/AuthEntityQueryMerchantCodeStatusRequest.json
@@ -1,0 +1,8 @@
+{
+    "appid": "ttxxxxxxxx",
+    "access_token": "xxxxxxxx",
+    "certificate_id": "100330xxxxxxx",
+    "industry_code": 10000,
+    "industry_role": 1,
+    "merchant_code": "99sxxxx"
+}
