+++ test/SKIT.FlurlHttpClient.ByteDance.MicroApp.UnitTests/ModelSamples/AuthEntity/AuthEntityQueryMerchantCodeStatusResponse.json
@@ -1,0 +1,8 @@
+{
+    "err": {
+        "err_code": 0,
+        "err_msg": ""
+    },
+    "status": "SUCCESS",
+    "status_desc": ""
+}
