+++ test/SKIT.FlurlHttpClient.ByteDance.MicroApp.UnitTests/ModelSamples/AuthEntity/AuthEntityRoleBindOrUnRequest.json
@@ -1,0 +1,7 @@
+{
+    "appid": "ttb7acc1112xxx277d01",
+    "access_token": "0801121846307a677xxx634338303473326e3573324c4d5779673d3d",
+    "merchant_entity_id": "E_71802667xxx55304249",
+    "industry_code": 10000,
+    "merchant_industry_role": 3
+}
