+++ test/SKIT.FlurlHttpClient.ByteDance.MicroApp.UnitTests/ModelSamples/AuthEntity/AuthEntityRoleBindOrUnResponse.json
@@ -1,0 +1,6 @@
+{
+    "err": {
+        "err_code": 0,
+        "err_msg": ""
+    }
+}
