+++ test/SKIT.FlurlHttpClient.ByteDance.MicroApp.UnitTests/ModelSamples/AuthEntity/AuthEntityUnAuthRoleRequest.json
@@ -1,0 +1,7 @@
+{
+    "appid": "tt5ed3286ec1exxxb801",
+    "access_token": "0801121846307a696b6143xxxb76736f356262425473753074773d3d",
+    "partner_entity_id": "E_720848123494xxx9309",
+    "merchant_entity_id": "E_72106048172xxx98392",
+    "merchant_industry_role": 2
+}
