+++ test/SKIT.FlurlHttpClient.ByteDance.MicroApp.UnitTests/ModelSamples/AuthEntity/AuthEntityUpdateAuthLetterResponse.json
@@ -1,0 +1,7 @@
+{
+    "err": {
+        "err_code": 0,
+        "err_msg": ""
+    },
+    "audit_taskid": "AppidBind_721021xxx2534894636"
+}
