+++ test/SKIT.FlurlHttpClient.ByteDance.MicroApp.UnitTests/ModelSamples/AuthEntity/AuthEntityUpdateBasicAuthResponse.json
@@ -1,0 +1,7 @@
+{
+    "basic_auth_taskid": "Basic_7179078906694500396",
+    "err": {
+        "err_code": 0,
+        "err_msg": ""
+    }
+}
