+++ test/SKIT.FlurlHttpClient.ByteDance.MicroApp.UnitTests/ModelSamples/AuthEntity/AuthEntityUpdateClassAuthResponse.json
@@ -1,0 +1,7 @@
+{
+    "class_auth_taskid": "Role_7179078906694500396",
+    "err": {
+        "err_code": 0,
+        "err_msg": ""
+    }
+}
