+++ test/SKIT.FlurlHttpClient.ByteDance.MicroApp.UnitTests/ModelSamples/AuthEntity/AuthEntityUploadMaterialResponse.json
@@ -1,0 +1,5 @@
+{
+    "err_no": 0,
+    "err_tips": "",
+    "path": "certification/resource/c60cbb3b8bb043cc9b349bf5484e1980"
+}
